vjtop warning规则应该是实际流逝时间，而不是设定刷新间隔计算 #107

+++ vjtop/src/main/java/com/vip/vjtools/vjtop/WarningRule.java
@@ -28,7 +28,7 @@ public class WarningRule {
 		thread.red = processors <= 8 ? processors * 225 : Math.max(8 * 225, processors * 150);
 	}
 
-	public void updateInterval(int intervalSeconds) {
+	public void updateInterval(long intervalSeconds) {
 		newThread.yellow = 1;
 		newThread.red = intervalSeconds * 2;
